Improved: Have a menu in Marketing featuring actions to create main objects (OFBIZ-11783)
Currently the create buttons for the main objects of the marketing application are located
within find and profile widgets/templates of those objects.

In order to improve the usability of this application, OFBiz,
and thus the appeal of it for adopters and users, these create buttons/links/etc.
should be in a main action menu visible at all times when a user is working within
the application of the component.

Modified:
MarketingMenus.xml - added MainActionMenu
To following screens - added ref to MainActionMenu, clean-up
CommonScreens.xml
ContactListScreens.xml
DataSourceScreens.xml
MarketingCampaignScreens.xml
MarketingReportScreens.xml
SegmentScreens.xml
TrackingCodeScreens.xml