Improved: Have a Menu in Humanres featuring actions to create the main objects (OFBIZ-11777)
Currently the create buttons for the main objects of the humanres plugin are located within find and profile widgets/templates of those objects.

In order to improve the usability of this plugin, OFBiz (and thus the appeal of it for adopters and users) these create buttons/links/etc. should be in a main action menu visible at all times when a user is working within the web appliclation of the component.

Modified:
HumanresMenus.xml - added MainActionMenu
To following screens - added ref to MainActionMenu, clean-up
CommonScreens.xml
EmplLeaveScreens.xml
EmployeeScreens.xml
EmploymentScreens.xml
PartyQualScreens.xml
PartyResumeScreens.xml
PartySkillsScreens.xml
PerfReviewScreens.xml