Fixed: CheckStyleMain Task is Failing in 18.12.04 (OFBIZ-12473)
Because of Windows EOL difference with *nix, despite the number for maxErrors
being more around 39600+ we temporarily put 50000 here to avoid issue when
releasing. Just increasing the number of lines is the problem.

The problem was already fixed in trunk by allowing all types of EOL, ie using
NewlineAtEndOfFile value="lf_cr_crlf"