Improved: Have a Menu in Catalog featuring actions to create the main objects (OFBIZ-11774)
Currently the create buttons for the main objects of the catalog components are located within the find and profile widgets/templates of those object.
In order to improve the usability of OFBiz (and thus the appeal of it for adopters and users) these create buttons/links/etc. should be in a main action menu visible at all times when a user is working within the component.

added: main action menu, having menu-items for
- new product process
- new store process
- new catalog process
- new category process
subject to catalog create permissions

added: decorator section for inclusion of main action menu where appropriate

Some cleanup