Improved: Move edit of Request to RequestSubTabBar (OFBIZ-12494)
In general, a <object>TabBar menu is intended to give users access to overviews related to the <object>. And a <object>SubTabBar menu is intended to give users access to actions to be performed on the <object>, like status changes, adding new related records, etc.
The menu-item for editing a request is currently located in the RequestTabBar menu in OrderMenus.xml.
Given that this menu-item is intended for users with UPDATE permissions and to give those users access to a screen to edit the data, this menu-item must move to the RequestSubTabBar menu.

modified: OrderMenus.xml
- menu RequestTabBar, removed menu-item editRequest (moved to RequestSubTabBar), additional cleanup
- menu RequestSubTabBar, added menu-item editRequest, added permission condition, additional cleanup