Improved: Routing (Task) Screens - consistency in ux (OFBIZ-12500)
Currently, when accessing screens related to production schemas and schema tasks (routing and routing tasks) in Manufacturing, the title of the shown object is below the menu items for that object (as shown in image attached).
To get to a more consistent user experience, this should be above the menu items for the object, like is done in many other screens in OFBiz.

modified: RoutingScreens.xml
CommonRoutingDecorator: moved title of object to above its menu, additional cleanup
CommonRoutingTaskDecorator: moved title of object to above its menu, additional cleanup