Improved: no functional change, labels improvements
In French, and maybe other languages, there is not only one word for new.
There are nouveau (masculine), nouvelle (féminine) and nouvel (masculine followed
by a word starting with a vowel)

Here are few related changes with misc. labels, and use of these labels in
MainActionMenu of marketing order and catalog menus