Improved: Move edit of Payment Group to SubTabBar (OFBIZ-12501)
In general, a <object>TabBar menu is intended to give
users access to overviews related to the <object>.
And a <object>SubTabBar menu is intended to give
users access to actions to be performed on the <object>,
like status changes, adding new related records, etc.
The menu-item for editing a payment group  is currently
located in the PaymentGroupTabBar menu in AccountingMenus.xml.
The menu item  belongs in the PaymentGroupSubTabBar

Improved: AccountingMenus.xml
PaymentGroupTabBar: removed menu item to edit the payment group, additonal cleanup
PaymentGroupSubTabBar: Added menu iteem to edit the payment group, additional cleanup