Improved: Use NPM with gradle to get external JS dependencies (OFBIZ-11960)
This replaces jquery-ui-timepicker-addon-1.6.3 by its equivalent from npm.

It should be noted that the 1.6.3 version is the EOL and
jquery-ui-timepicker-addon is no longer maintained.

Despite that according to
https://snyk.io/advisor/npm-package/jquery-ui-timepicker-addon
it's still a good choice, notably there are known no security issues.

I tried to use https://www.npmjs.com/package/jquery-datetimepicker instead
but crossed too much issues and let it go.