Improved: Quote Search-Find - Consistency in UX (OFBIZ-12506)
Throughout the various components the screens to find a particular record (or multiple) are structured the same way with a decorator-screen called 'FindScreenDecorator' and in there decorator section for 'search-options' and 'search-results', bringing to the users a consistent experience.
The search and find functionality for quotes does not adhere to that paradigm.

modified: QuoteScreens.xml
restructured FindQuote screen to align with paradigm applied to the same vis-a-vis
searching and finding other kinds of records, additional cleanup.