Improved: CommonRequestDecorator - Consistency (OFBIZ-12507)
The CommonRequestDecorator screen has the title and
the action menu for users with CREATE and/or UPDATE
permissions in the decorator section for the body
of a request screen.
To improve consistency across similar screens for
other objects, the elements should move to the
decorator section for the pre-body of request screens.

modified: CommonScreens.xml
moved RequestSubTabBar and title for a request
from decorator-section body to decorator-section pre-body,
additional cleanup