Improved: Request - Move new item action to menu (OFBIZ-12508)
Currently the action trigger for adding a new item to
a Request, by users with CREATE permissions is l
ocated in screens related to a Request.
This action trigger should be located in the
Action Menu for the request.

modified:
- CustRequestScreens.xml - removed action trigger(s) and surrounding style elements
from screens, additional cleanup
- OrderMenus.xml - adding NewRequestItem menu-item for users with
CREATE permissions to add new request items to a request.