Improved: CommonStoreDecorator - consistency (OFBIZ-12537)
The common decoratr screen for store records has the title and the action menu for users with CREATE and/or UPDATE permissions in places inconsistent to similar screens for other objects. The elements should move to improve the consistency of the user experience.

modified: CommonScreens.xml
relocated title and the object action menu for a store record.
additional cleanup