Fixed: Upgrade Tomcat from 9.0.54 to 9.0.58 (OFBIZ-12539)
The fix for bug CVE-2020-9484 introduced a time of check, time of use
vulnerability that allowed a local attacker to perform actions with the
privileges of the user that the Tomcat process is using. This issue is only
exploitable when Tomcat is configured to persist sessions using the FileStore.

Note: because I'm a kind of outlaw (using deprecated Win 7) I must use node
13.14.0 and it causes me some issues, like the previous revert.
You can't have your cake and eat it