Improved: Have a Menu in Content featuring actions to create the main objects (OFBIZ-12541)
Currently the create buttons for the main objects of the Content application are located within find and profile widgets/templates of those objects.
In order to improve the usability of this application, OFBiz and thus the appeal of it for adopters and users, these create buttons/links/etc. should be in a main action menu visible at all times when a user with CREATE permissions is working within the application

modified:
ContentMenuxs.xml - added MainActionMenu for users with CREATE permissions in the Content application
Various screens in the content component - added MainActionMenu where appropriate
additional cleanup