Fixed: Wrong Field Name Definition in RequirementForms (OFBIZ-12505)
The issue exists also in the older version.
The field "requirementByDate" form "FindApprovedProductRequirements" in RequirementForms.xml has been set wrongly and it should be "requiredByDate".
This causes the result could not be queried out properly as indicated as below:
The result is governed by the service "getRequirementsForSupplier", where reads the list from the view "RequirementAndRole".
In the view entity definition, the field is defined as
<alias entity-alias="RQ" name="requiredByDate"/>, but it is not the case in the form "FindApprovedProductRequirements" above.

Modified: RequirementForms.xml
- changed field requirementByDate into requiredByDate as per entity definition for Requirement