Fixed: Element condition-[expr/list/object] doesn't work with compound (OFBIZ-12546)
When you create an entity-condition in a form or grid under a compound element, to write a valid xml you write it with the namespace like this :

    <wf:actions>
        <wf:entity-condition entity-name="PartyRelationship"...>
            <wf:condition-list>
                <wf:condition-expr field-name="partyIdFrom" from-field="parameters.partyIdFrom"/>
                <wf:condition-expr field-name="partyIdTo" from-field="parameters.partyIdTo"/>
            </wf:condition-list>
        </wf:entity-condition>
    </wf:actions>

This failed when OFBiz execute the search with error Invalid element with name '[wf:condition-expr]' found under a condition-list element.

The error came from the class EntityFinderUtil.ConditionList who analyze the condition-list's children with the namespace + node name instead take only the node name.

    if ("condition-expr".equals(subElement.getNodeName()))

To solve this issue we only check the local node name regardless from the namespace used.