Fixed: Manufacturing Jobshop find screen by default does not show all production runs (OFBIZ-12550)
The manufacturing jobshop find screen has the seach date initialized to nowdate.
if one wants to see all production runs, one has to clear the date which should
be empty initially

The same problem exists on the MRP log find where the date is filled by default.

Thanks: Hans for report and patches