Fixed: GlSetupScreens.xml - MainActionMenu not shown (OFBIZ-12559)
The MainActionMenu of the Accounting component is
intended to provide the users with CREATE permissions
a direct way to create the main objects of the
Accounting components (Gl transaction, invoice,
payment), instead of - as such a user - have to go
through multiple screens to get to the action
trigger to create such objects.
The MainActionMenu is applied in various decorator
screens. It is, however, not applied in the
GlSetupScreen.xml file.

Modified: GlSetupScreens.xml
added pre-body decorator section, including ref to
MainActionMenu, to:
- screen ListCompanies
- screen AddCompany
- screen ImportExport