Fixed: Party Main: duplicate action trigger (Person/Group) (OFBIZ-12560)
When accessing
https://localhost:8443/partymgr/control/main the
screen shows duplicate action triggers for the
creation of a new party of type=PERSON or
of type=PARTYGROUP.
These action triggers come from the MainActionMenu
for the a component, and from the 'CreateNewParty'
menu.

modified: PartyMenus.xml
removed from menu CreateNewParty:
- menu-item for the creation of a PartyGroup
- menu-item for the creation of a Person
as these already exists in the MainActionMenu of
the component.