Documented: Continue the common-theme upload (OFBIZ-9138 Create a common theme)
Removes 2 FIXMEs about locations for image.server.path and image.management.path
properties

This follows a message I sent to the user ML
https://lists.apache.org/thread/syysbhl99h0753h77hzd5frb50foclsq
that is:
<<Actually I was wrong when I said that "a feature was lost when common-theme was
put in". The rest is right. This feature is the possibility, through
image.server.path property in catalog.properties file, to place the images, and
other the static files as well, in a location that fits with you for
any reason. Notably following the NSA recommendation to place it in "a non-web
accessible area". This to prevent webshell uploads and all kind of
other malicious files uploads. The same is true for the other property
image.management.path.

So the fact that before common-theme was put in, with the folder for images
/themes/common/images/webapp/images/, this folder was
/framework/images/webapp/images/ has nothing to do with "a non-web accessible
area". That's you to decide...

There is also a ${tenantId} var used in image.server.path property that is used
in case of multi-tenant, that's another thing.

So I finally don't think it's necessary to put the images and image.management
in runtime. This would add nothing. I'll remove the FIXMEs>>

Thanks: APL+Win to help me format the message to 80 chars in a breeze :)