Improved: Prevent possible DOS attack done using Java deserialisation (OFBIZ-12592)
Qing Xu, a security reporter, alerted us that, despite no current vulnerability,
it could be maybe possible to do DOS attacks using Java deserialisation.

That has been fixed with https://openjdk.java.net/jeps/290 and even implemented
in Java 8, but it needs a little effort on our side, here it is.

There is no real bug, to this is an improvement but a security improvement so it
will be backported

Thanks: Qing Xu for report

Conflicts handled by hand gradle.properties