Improved: no functional change, adds GitHub Actions build status emails.
As the name suggests, these are emails that notify of either a failed build
using GitHub Actions, or a change from failure in previous builds to a
successful build. Continuously successful builds will not trigger emails, as
that could be rather spammy.

Please see the documentation on this at: https://s.apache.org/asfyaml-gha