Fixed: inv. reservation places order if not enough qty (OFBIZ-12455)
If quantity not reserved is not 0, requireInventory is set to Y, system
let an order to be placed anyway.

The usage of the service around the project let make an important
assertion: we can consider that also reserveInventory flag of the
ProductStore is Y, since right now is always the caller context that
checks for the reserveInventory flag.

This fix simply returns a specific error to the caller, avoiding the
system to place the order.

Thanks Nicola Mazzoni for helping in debug this issue and Jacopo
Cappellato for the patch review.