Fixed: ecomm. listing ProductFacility out of stock issue (OFBIZ-12359)
In method filterOutOfStockProducts added case for "virtual" products.
In case a product isVirtual (= Y), then the available inventory count
(lastInventoryCount) is summed up on all its product variants, across
all the ProductFacility records, and then used to determine if the
(parent/virtual) product is in stock or out of stock.

This fix does not check if the facilityId of the ProductFacility
records retrieved is associated and enabled to the ProductStore.

Thanks Nicola Mazzoni for helping in anlyzing this issue.