Improved: GH Actions and Buildbot Builders are not consistent (INFRA-23076)
It's now 2 weeks that INFRA-23076 is open, so like 3 weeks that BuildBot is not
working for OFBiz at all.

So far I did not run the integration tests on GH because
* we have it on BuildBot
* it's complicated to show the results

Hopefully BuildBot will soon run the OFBiz builds again. But it does not hurt
to run the integration tests on GH, even w/o the results, because at least it's
easier than running them for each commit locally (takes time) and we can run
them locally once we know there is a problem demonstrated with GH actions.