Fixed: RequestMap.event with type service multi doesn't work with dynamic url (OFBIZ-12604)
When you define a controller request-map with an event of type 'service-multi', if your uri contains multiple allocation like 'MyWay/MyAction' the class ServiceMultiEventHandler failed to execute with an EventHandlerException

 *****
    <request-map uri="Payment/QuickSend">
        ...
        <event type="service-multi" invoke="quickSendPayment"/>
        ...
 *****

The reason comes from the necessary to resolve the attribute global-transaction on event definition, and to do that a call to ConfigXMLReader is realized.

But unnecessary because we already have the event element on the context.
Use it directly, simplify the code and fix this issue