Fixed: Content tag in a screen does not display correctly images (OFBIZ-12685)
When we want to display an image content within a screen through <content/> tag,
images are not rendered correctly
To test:
Go to admin party profile page and upload a picture file as a new "LGOIMGURL"
for instance: https://localhost:8443/partymgr/control/viewprofile?partyId=admin

When your content file is uploaded, note the contentId and got to page
ShowContent (which render content using the tag <content />)
https://localhost:8443/content/control/showContent?contentId=XXXXX
The image will not be rendered without the given patch.

jleroux: since it's eventually negated, it makes no sense to test if the
content is either an image or an application. It can't be both. This is a very
old feature (pre-apache era) that seems to have never been tested/used
I have made a formatting change and added a comment to explain.

Thanks: Leila Mekika