Improved: Add auto-parameters-form option to on-event-update-area link in xml form (OFBIZ-12684)
As we have an auto-parameters-entity and an auto-parameters-service option in
xml forms on-event-update-area tags, it could be useful to have an
auto-parameters-form option that will propagate current form fields as parameters
to the paginate or submit action link (for instance in modal "tunnel" processes)

Currently, if you want to propagate the parameters of a form to a zone that is
refreshed in ajax after the submit you have to write specific javascript code or
to return all the parameters sent to the request through a service as OUT
parameters and/or redirect-parameter.

This feature will add automatically the form parameters on the
generated on-event-update-area link

jleroux: fixed 3 checkstyle issues:
> Task :checkstyleMain
[ant:checkstyle] [ERROR] CommonWidgetModels.java:385:17:
Il y a une espace de trop après '!'. [NoWhitespaceAfter]

[ant:checkstyle] [ERROR] CommonWidgetModels.java:398:77:
Il manque une espace après 'conversion de type'. [WhitespaceAfter]

[ant:checkstyle] [ERROR] ModelForm.java:2326:
La ligne excède 150 caractères (trouvé 155). [LineLength]

Thanks: Leila