Improved: The big problem when loading seed. (OFBIZ-7754)
Regarding, OFBIZ-7112,
that's good for who start on use the ofbiz with initial setup, but not for the
site that already online and has to update the OFBiz core. Because when has
update OFBiz core they will use command load-seed for update.

The problem is if we use load-seed mean the configuration data that's already
exists will be replaced by the data from this file, CommonSystemPropertyData.xml

So, for my suggestion should change the reader from seed to seed-initial or
remove systemPropertyValue from the data file.

Thanks: Kongrath Suankaewmanee