Reverted: "Fixed: Extend HTML Sanitizer - style attribute (OFBIZ-12691)"
This reverts commit 02387ffea720dd5f45b2abb34dc61e0f39868d97.

I'm not sure why when coming from UI HtmlSanitizer.Policy() change quotes
to HTML entities, but not when coming from test. It seems to come from the
context as org.owasp.html.HtmlStreamRenderer somehow explains it:

<<Given a series of HTML tokens, writes valid, normalized HTML to the output.
The output will have well-defined tag boundaries, but there may be orphaned or
missing close and open tags. The result of two renderers can always be
concatenated to produce a larger snippet of HTML, but if the first
was called with writeOpenTag("plaintext", ...), then any tags in the second will
not be interpreted as tags in the concatenated version.>>

Anyway reverting fixes the test issue.