Fixed: Unable to upload a CSV file (OFBIZ-12636)
While testing on "Upgrade pdfbox to 2.0.27 (OFBIZ-12698)" I found that I could
not upload a valid PDF.

Actually it was due to SecuredUpload::isValidCsvFile that was crashing when
uploading this PDF file.

I don't remember clearly but it was for a reason that I did not put the call to
isValidCsvFile as the last check in isValidFile for the "all" and default case.

As I can't remember now and it works for both CSV and PDF as "all" (when adding
files from an user profile) I eventually decided to put "isValidCsvFile" at the
end of the check. Unfortunately, it's complicate to test all possible combinations.