Improved: Reduced use of temporary variables when rendering date-time fields (OFBIZ-12126)
For cases where render properties of the date-time field are calculated
simply, apply the result directly to the RenderableFtlMacroCallBuilder
rather than saving the results in temporary variables and assigning at
the end of the method.

The aim of this change is to make code easier to read as the developer
does not need to keep track of so many variables.