Improved: MacroFormRenderer refactoring of datetime fields (OFBIZ-12126)
Part of the OFBIZ-11456 MacroFormRenderer refactoring effort.

Rather than MacroFormRenderer producing and evaulating FTL strings, it now uses RenderableFtlElementsBuilder to create RenderableFtlMacroCall objects for datetime fields which are then passed to an FtlWriter for evaluation.

Added tests to document how rendering parameters are generated for datetime fields.

Simplified generation of selectable minute options for date-time fields used in time-dropdown mode.

Moved typing of the date-time field's step, type, mask and clock attributes to DateTimeField model to ensure they are checked when the model is instantiated. The type and clock attributes are exposed by the model as boolean getters to check field types and whether a 12 or 24-hour clock should be used. These changes remove the need for client code to decode the various string attributes.

Removed redundant code which enabled/disabled encoding of date-time field output