Fixed: newsletter multiple subscript. w same email address (OFBIZ-9361)
(backport)
It was possible to be subscribed to the newsletter with the same email
address multiple time.

Added a check for already ACCEPTED requests in order to refuse new
incoming subscription with the same email address.
For existing PENDING requests instead, has been added the deletion of
all the existing records for the same email address, before adding a
new one.

Thanks Aditya Sharma for your feedback on this.