FIXED: Added mandatory seed data needed for Purchase Order creation (OFBIZ-12608)
When creating an order it is necessary to specify shipping options
and therefore ensure that entities representing those shipping options
already exist in the OFBiz database. The ofbizsetup application imports
some default shipping options when creating a Product Store.

In the case of purchase orders, shipping is handled by the supplier, but
the OFBiz data model still requires that shipping information is
recorded against the order. In this case OFBiz uses some hard-coded
shipping options, but the corresponding entities for those shipping
options were not included in seed data.

This fix moves the entities for the hard-coded purchase order shipping
options to seed data.